// charts/types.ts
export type SessionData = {
    m_weather: number;
    m_trackTemperature: number;
    m_airTemperature: number;
    m_totalLaps: number;
    m_trackLength: number;
    m_sessionType: number;
    m_trackId: number;
    m_formula: number;
    m_sessionTimeLeft: number;
    m_sessionDuration: number;
    m_aiDifficulty: number;
    m_pitStopWindowIdealLap: number;
    m_pitStopWindowLatestLap: number;
    m_pitStopRejoinPosition: number;
    m_timeOfDay: number;
  };